version 1.0.2 update
1. [New feature] new function writeConnectivityMap2VTK() is added for Python interface which allows to output pts connectivity map into vtk
2. [Perfromance] memory usage is optimized by pass-by-reference code style
3. [Example] example code is added to reorder gmsh files using meshio

+++ visual_studio/hilbertSort.h
@@ -276,7 +276,7 @@ inline void Zoltan_quicksort_pointer_inc_double(
 )
 {
     int i, j;
-    double temp;
+    int temp;
 
     /* sort in decreasing order */
     Zoltan_quicksort_pointer_dec_double(sorted, val, start, end);
